chapter9/firestore/requirements.txt
Flask
google-cloud-storage
google-cloud-firestore
